grouped gemm: accept canonical (sum_m,k)/(l,n,k) layouts and flat SF buffers
The contiguous grouped GEMM SwiGLU/dSwiGLU wrappers and API classes now
additionally accept natural row-major inputs -- A (sum_m, k), B (l, n, k)
C-contiguous, dense C-contiguous SFA/SFB buffers (flat or physical atom
shape), prob (sum_m,) fp32/bf16 -- and normalize them internally. Canonical
SF buffers compile as flat 1-D pointers: the kernels already rebuild the
MMA-tiled SF layouts from the GEMM shapes and read only the base pointer.
Canonical calls return natural-shaped outputs; alpha defaults to cached
ones. Pre-permuted kernel-facing inputs keep working unchanged.

+++ python/cudnn/gemm/cutedsl/grouped/dswiglu/grouped_gemm_dswiglu_quant.py
@@ -2580,7 +2580,7 @@ class BlockScaledContiguousGroupedGemmKernel:
                 # Note, it always assumes T2R_M/EPI_M is 1, otherwise it will break the result.
                 #
                 mPosition = tile_info[0] * self.mma_tiler[0] // cute.size(tiled_mma.thr_id.shape) + tidx
-                mProb = prob[mPosition, 0, 0]
+                mProb = prob[mPosition, 0, 0].to(cutlass.Float32)
                 if cutlass.const_expr(self.generate_dprob):
                     dProbVal = cutlass.Float32(0.0)
